backend/requirements.txt
fastapi
uvicorn[standard]
cryptography
pyotp
python-multipart
passlib[argon2]
sqlalchemy
pydantic-settings
itsdangerous
alembic